// test-day/src/app/feature/weeks-schedule/models/medicines.ts
export interface Medicines {
    _id: number;
    name: string;
    time: Date;
    dose: string
    day: WeekDays;
}

export enum WeekDays {
    Monday = 'Monday',
    Tuesday = 'Tuesday',
    Wednesday = 'Wednesday',
    Thursday = 'Thursday',
    Friday = 'Friday',
    Saturday = 'Saturday',
    Sunday = 'Sunday'
}